+++ frontend/src/components/PermissionConfig/ProjectPermissionDashboard.js
@@ -17,34 +17,51 @@ const ProjectPermissionDashboard = () => {
   const [allProjects, setAllProjects] = useState({});
   const [tableHeaderData, setTableHeaderData] = useState([]);
 
-  const [allAvailableProjects, setAllAvailableProjects] = useState({});
+  const [allUsersPrivateProjects, setAllUsersPrivateProjects] = useState({});
   const [tableBodyData, setTableBodyData] = useState([]);
 
   /*
-    Fetching All projects we have from Project API which becomes a table's header
+    Fetching All projects(Public and Private) we have from Project table which becomes a table's header
   */
   useEffect(() => {
     const abortController = new AbortController();
     const signal = abortController.signal;
 
-    const getProjects = async () => {
+    const getAllProjects = async () => {
       try {
         const token = await getTokenSilently();
 
-        await fetch(
-          CONSTANTS.CORE_API_BASE_URL +
-            CONSTANTS.INTERMEDIATE_API_ALL_PROJECTS_ENDPOINT,
-          {
-            headers: {
-              Authorization: `Bearer ${token}`,
-            },
-            signal: signal,
-          }
-        )
+        await Promise.all([
+          fetch(
+            CONSTANTS.CORE_API_BASE_URL +
+              CONSTANTS.INTERMEDIATE_API_ALL_PUBLIC_PROJECTS_ENDPOINT,
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              },
+              signal: signal,
+            }
+          ),
+          fetch(
+            CONSTANTS.CORE_API_BASE_URL +
+              CONSTANTS.INTERMEDIATE_API_ALL_PRIVATE_PROJECTS_ENDPOINT,
+            {
+              headers: {
+                Authorization: `Bearer ${token}`,
+              },
+              signal: signal,
+            }
+          ),
+        ])
           .then(handleErrors)
-          .then(async (response) => {
-            const responseData = await response.json();
-            setAllProjects(responseData);
+          .then(async ([publicProjects, privateProjects]) => {
+            const publicProjectsData = await publicProjects.json();
+            const privateProjectsData = await privateProjects.json();
+
+            setAllProjects({
+              public: { ...publicProjectsData },
+              private: { ...privateProjectsData },
+            });
           })
           .catch((error) => {
             console.log(error);
@@ -54,7 +71,7 @@ const ProjectPermissionDashboard = () => {
       }
     };
 
-    getProjects();
+    getAllProjects();
 
     return () => {
       abortController.abort();
@@ -63,7 +80,7 @@ const ProjectPermissionDashboard = () => {
 
   /*
     Based on all projects name we pulled from above useEffect Hook,
-    Create an array of objects to set the table's header
+    Create an readable array of objects for material-ui table
   */
   useEffect(() => {
     // If and only if the object is not empty, create list for table's header
@@ -78,24 +95,30 @@ const ProjectPermissionDashboard = () => {
           label: "Auth0 ID",
         },
       ];
-      for (const [key, value] of Object.entries(allProjects)) {
-        tempArray.push({
-          id: key,
-          label: value.project_full_name,
-        });
+
+      for (const access_level of Object.keys(allProjects)) {
+        for (const [project_id, project_name] of Object.entries(
+          allProjects[access_level]
+        )) {
+          tempArray.push({
+            id: project_id,
+            label: `${project_name} - ${access_level}`,
+          });
+        }
       }
       setTableHeaderData(tempArray);
     }
   }, [allProjects]);
 
   /*
-    Pull every row from Available_Project table
+    Pull every row from Users_Projects table
+    Which is Private Project
   */
   useEffect(() => {
     const abortController = new AbortController();
     const signal = abortController.signal;
 
-    const getAllAvailableProjects = async () => {
+    const getAllUsersProjects = async () => {
       try {
         const token = await getTokenSilently();
 
@@ -112,7 +135,7 @@ const ProjectPermissionDashboard = () => {
           .then(handleErrors)
           .then(async (response) => {
             const responseData = await response.json();
-            setAllAvailableProjects(responseData);
+            setAllUsersPrivateProjects(responseData);
           })
           .catch((error) => {
             console.log(error);
@@ -122,7 +145,7 @@ const ProjectPermissionDashboard = () => {
       }
     };
 
-    getAllAvailableProjects();
+    getAllUsersProjects();
 
     return () => {
       abortController.abort();
@@ -132,15 +155,15 @@ const ProjectPermissionDashboard = () => {
   /*
     Create an array of objects for table body
     We need to be sure that the following data aren't empty
-    1. allAvailableProjects -> Data from UserDB, Available_Project table
-    2. allProjects -> Data from Project API, All Projects we provide
+    1. allUsersPrivateProjects -> Data from Users_Projects table
+    2. allProjects -> Data from Project table, All Public & Private projects
     3. userOption -> Data from Auth0, existing users
   */
   useEffect(() => {
     if (
-      (allAvailableProjects &&
-        Object.keys(allAvailableProjects).length === 0 &&
-        allAvailableProjects.constructor === Object) === false &&
+      (allUsersPrivateProjects &&
+        Object.keys(allUsersPrivateProjects).length === 0 &&
+        allUsersPrivateProjects.constructor === Object) === false &&
       (allProjects &&
         Object.keys(allProjects).length === 0 &&
         allProjects.constructor === Object) === false &&
@@ -150,8 +173,8 @@ const ProjectPermissionDashboard = () => {
       let tempObj = {};
 
       /*
-        Loop through the object, allAvailableProjects
-        Nested loop through another object, allprojects
+        Loop through the object, allUsersPrivateProjects
+        Nested loop through another object, allProjects
 
         The key for a temp object property with "auth0-user-id", 
         is user-email. (using find function to find an object with auth0-id value
@@ -159,24 +182,30 @@ const ProjectPermissionDashboard = () => {
         {value: auth0-id, label: email | auth0 or google auth}
         Then from a found object, use the label as we want email to be displayed
 
-        Then, compare if user's available_projects (projects with permission),
+        Then, compare if user's user_private_projects (projects with permission),
         contains the project we provide, then it's true else false.
         )
       */
-      for (const [user_id, available_projects] of Object.entries(
-        allAvailableProjects
+      for (const [user_id, user_private_projects] of Object.entries(
+        allUsersPrivateProjects
       )) {
-        for (const [project_code, project_detail] of Object.entries(
-          allProjects
-        )) {
+        for (const access_level in allProjects) {
           tempObj["auth0-user-id"] = userOption.find(
             (user) => user.value === user_id
           ).label;
-          tempObj[project_code] = available_projects.includes(
-            project_detail.project_id
-          )
-            ? "true"
-            : "false";
+          // If access_level is public, default to true, means have permission to use
+          if (access_level === "public") {
+            for (const project_id in allProjects[access_level]) {
+              tempObj[project_id] = "true";
+            }
+            // acces_level is not public, then compare with Users_Projects to check the permission
+          } else {
+            for (const project_id in allProjects[access_level]) {
+              tempObj[project_id] = user_private_projects.includes(project_id)
+                ? "true"
+                : "false";
+            }
+          }
         }
         tempArray.push(tempObj);
         tempObj = {};
@@ -184,7 +213,7 @@ const ProjectPermissionDashboard = () => {
 
       setTableBodyData(tempArray);
     }
-  }, [allAvailableProjects, allProjects, userOption]);
+  }, [allUsersPrivateProjects, allProjects, userOption]);
 
   /*
     Fetching user information from the API(Auth0)