+++ middleware/middleware/db.py
@@ -15,8 +15,8 @@ def _is_user_in_db(user_id):
     return bool(models.User.query.filter_by(user_id=user_id).first())
 
 
-def _add_user_to_db(user_id):
-    """Add an user to the MariaDB if not exist
+def _insert_user(user_id):
+    """Insert an user to the MariaDB if not exist
 
     Parameters
     ----------
@@ -31,33 +31,33 @@ def _add_user_to_db(user_id):
         print(f"User {user_id} already exists")
 
 
-def _is_project_in_db(project_name):
+def _is_project_in_db(project_id):
     """To check whether the given project is in the DB
 
     Parameters
     ----------
-    project_name: string
+    project_id: string
         A project code we use internally.
         E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
     """
-    return bool(models.Project.query.filter_by(project_name=project_name).first())
+    return bool(models.Project.query.filter_by(project_id=project_id).first())
 
 
-def _add_project_to_db(project_name):
-    """Add a new project to the MariaDB if not exist
+def _insert_project(project_id):
+    """Insert a new project to the MariaDB if not exist
 
     Parameters
     ----------
-    project_name: string
+    project_id: string
         A project code we use internally.
         E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
     """
-    if not _is_project_in_db(project_name):
-        db.session.add(models.Project(project_name))
+    if not _is_project_in_db(project_id):
+        db.session.add(models.Project(project_id))
         db.session.commit()
         db.session.flush()
     else:
-        print(f"Project {project_name} already exists")
+        print(f"Project {project_id} already exists")
 
 
 def _is_permission_in_db(permission_name):
@@ -74,8 +74,8 @@ def _is_permission_in_db(permission_name):
     )
 
 
-def _add_permission_to_db(permission_name):
-    """Add new permission to the MariaDB if not exist
+def _insert_permission(permission_name):
+    """Insert new permission to the MariaDB if not exist
 
     Parameters
     ----------
@@ -91,7 +91,7 @@ def _add_permission_to_db(permission_name):
         print(f"Project {permission_name} already exists")
 
 
-def _is_user_permission_in_db(user_id, permission):
+def _is_user_access_permission_in_db(user_id, permission):
     """Check whether there is a row with given user_id & permission"""
     return bool(
         models.UserPermission.query.filter_by(user_id=user_id)
@@ -100,9 +100,9 @@ def _is_user_permission_in_db(user_id, permission):
     )
 
 
-def _add_user_permission_to_db(user_id, permission):
+def _insert_user_access_permission(user_id, permission):
     """Insert data(page access permission) to the bridging table,
-    allowed_permission
+    users_permissions
 
     Parameters
     ----------
@@ -117,10 +117,10 @@ def _add_user_permission_to_db(user_id, permission):
     )
     if not _is_permission_in_db(permission):
         print(f"{permission} is not in the DB so updating it.")
-        _add_permission_to_db(permission)
+        _insert_permission(permission)
         db.session.flush()
 
-    if not _is_user_permission_in_db(user_id, permission):
+    if not _is_user_access_permission_in_db(user_id, permission):
         db.session.add(models.UserPermission(user_id, permission))
         db.session.commit()
         db.session.flush()
@@ -128,65 +128,33 @@ def _add_user_permission_to_db(user_id, permission):
         print(f"{user_id} already has a permission with ${permission}")
 
 
-def _add_user_project_to_db(user_id, project_name):
-    """Insert data(allowed projects) to the bridging table,
-    allowed_projects
+def update_user_access_permission(user_id, permission_list):
+    """Update/Insert user's assigned permission to a table,
+    Users_Permissions
 
     Parameters
     ----------
     user_id: string
-        Selected user's Auth0 id
-    project_name: string
-        Selected project's project code.
-        E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
+        Auth0's unique user id
+    permission_list: list
+        List of permission that the user has. (From Auth0, trusted source)
     """
+    # Sync the users_permissions table to token's permission (trusted source)
+    # first before we check/update
+    _sync_permissions(user_id, permission_list)
 
-    print(f"Check whether the project is in the DB, if not, add the project to the DB")
-    if not _is_project_in_db(project_name):
-        print(f"{project_name} is not in the DB so updating it.")
-        _add_project_to_db(project_name)
+    print(f"Check whether the user is in the DB, if not, add the person to the DB")
+    if not _is_user_in_db(user_id):
+        print(f"{user_id} is not in the DB so updating it.")
+        _insert_user(user_id)
         db.session.flush()
 
-    # Find Find a project object with a given project_name to get its project id
-    project_obj = models.Project.query.filter_by(project_name=project_name).first()
-
-    db.session.add(models.UserProject(user_id, project_obj.project_id))
-    db.session.commit()
-    db.session.flush()
+    for permission in permission_list:
+        _insert_user_access_permission(user_id, permission)
 
 
-def _remove_user_projects_from_db(user_id, project_name):
-    """Remove data(project) from the bridging table,
-    allowed_projects
-
-    Parameters
-    ----------
-    user_id: string
-        Selected user's Auth0 id
-    project_name: string
-        Selected project's project code.
-        E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
-    """
-    try:
-        # Get the project id with the given project name
-        certain_project_id = (
-            models.Project.query.filter_by(project_name=project_name).first().project_id
-        )
-        allowed_projects_row = (
-            models.UserProject.query.filter_by(user_id=user_id)
-            .filter_by(project_id=certain_project_id)
-            .first()
-        )
-
-        db.session.delete(allowed_projects_row)
-        db.session.commit()
-        db.session.flush()
-    except:
-        print("Something went wrong.")
-
-
-def _remove_user_permission_from_db(user_id, permission):
-    """allowed_permission table is outdated, remove illegal permission to update the dashboard"""
+def _remove_user_access_permission(user_id, permission):
+    """users_permissions table is outdated, remove illegal permission to update the dashboard"""
     illegal_permission_row = (
         models.UserPermission.query.filter_by(user_id=user_id)
         .filter_by(permission_name=permission)
@@ -198,8 +166,83 @@ def _remove_user_permission_from_db(user_id, permission):
     db.session.flush()
 
 
-def get_user_projects(user_id):
-    """Retrieves all projects ids for the specified user
+def _insert_user_project_permission(user_id, project_id):
+    """Insert data(assigned projects) to the bridging table,
+    users_projects
+
+    Parameters
+    ----------
+    user_id: string
+        Selected user's Auth0 id
+    project_id: string
+        Selected project's project code.
+        E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
+    """
+
+    print(f"Check whether the project is in the DB, if not, add the project to the DB")
+    if not _is_project_in_db(project_id):
+        print(f"{project_id} is not in the DB so updating it.")
+        _insert_project(project_id)
+        db.session.flush()
+
+    # Find a project object with a given project_id to get its project id
+    project_obj = models.Project.query.filter_by(project_id=project_id).first()
+
+    db.session.add(models.UserProject(user_id, project_obj.project_id))
+    db.session.commit()
+    db.session.flush()
+
+
+def _remove_user_project_permission(user_id, project_id):
+    """Remove data(project) from the bridging table,
+    users_projects
+
+    Parameters
+    ----------
+    user_id: string
+        Selected user's Auth0 id
+    project_id: string
+        Selected project's project code.
+        E.g., gnzl, mac_raes, nzgs_pga, soffitel_qtwn...
+    """
+    try:
+        # Get the project id with the given project name
+        certain_project_id = (
+            models.Project.query.filter_by(project_id=project_id).first().project_id
+        )
+
+        users_projects_row = (
+            models.UserProject.query.filter_by(user_id=user_id)
+            .filter_by(project_id=certain_project_id)
+            .first()
+        )
+
+        db.session.delete(users_projects_row)
+        db.session.commit()
+        db.session.flush()
+    except:
+        print("Something went wrong.")
+
+
+def _get_user_access_permission(requested_user_id):
+    """Retrieve all permissions for the specified user
+
+    Returns
+    -------
+    A list of permission names
+    """
+    all_users_permissions_for_a_user_list = models.UserPermission.query.filter_by(
+        user_id=requested_user_id
+    ).all()
+
+    return [
+        permission.permission_name
+        for permission in all_users_permissions_for_a_user_list
+    ]
+
+
+def get_user_project_permission(user_id):
+    """Retrieves all projects ids for the specified user from Users_Permissions table
 
     Parameters
     ----------
@@ -208,124 +251,41 @@ def get_user_projects(user_id):
 
     Returns
     -------
-    list of Project IDs from DB (Allowed Projects)
+    list of Project IDs from DB (Assigned Projects from Users_Projects table)
     """
-    # Get all allowed projects that are allocated to this user.
-    allowed_project_objs = (
+    # Get all projects that are assigned to this user.
+    assigned_project_objs = (
         models.Project.query.join(models.UserProject)
         .filter((models.UserProject.user_id == user_id))
         .all()
     )
 
-    return [project.project_name for project in allowed_project_objs]
+    return {
+        project.project_id: project.project_name for project in assigned_project_objs
+    }
 
 
-def get_all_users_projects():
-    """Retrieve all allowed projects from Allowed_Project table
+def get_all_users_project_permissions():
+    """Retrieve all assigned projects from Users_Projects table for each user
 
     Returns
     -------
-    allowed_projects_dict: dictionary
+    users_projects_dict: dictionary
         In the form of:
         {
            userA: [ProjectA, ProjectB, ProjectC],
            userB: [ProjectA, ProjectB]
         }
     """
-    # Get all allowed projects from the UserDB
-    allowed_projects = models.UserProject.query.all()
+    # Get all assigned projects from the UserDB
+    users_projects = models.UserProject.query.all()
 
-    allowed_projects_dict = defaultdict(list)
+    users_projects_dict = defaultdict(list)
 
-    for project in allowed_projects:
-        allowed_projects_dict[project.user_id].append(project.project_id)
+    for project in users_projects:
+        users_projects_dict[project.user_id].append(project.project_id)
 
-    return allowed_projects_dict
-
-
-def get_all_projects_for_dashboard(projects):
-    """Retrieve all the projects we have from Project API
-
-    Then customize the format to send to the frontend for two reasons.
-    1. For Table's header(will display the name of the project)
-    2. We can use this dictionary to filter the table to tell they have permission
-
-    Parameters
-    ----------
-    projects: dictionary
-
-    Returns
-    -------
-    dictionary:
-        In the form of
-        {
-            project_id: project_full_name
-        }
-        project_id = (e.g., nzgl, soffitel,qtwn)
-        project_fulle_name = user friendly name for project,
-        e.g. Generic New Zealand Locations
-    """
-    # Get all projects from the UserDB.
-    all_projects = models.Project.query.all()
-
-    return {
-        project.project_name: {
-            "project_id": project.project_id,
-            "project_full_name": projects[project.project_name]["name"],
-        }
-        for project in all_projects
-    }
-
-
-def allocate_projects_to_user(user_id, project_list):
-    """Give user a permission of the chosen projects
-
-    Parameters
-    ----------
-    user_id: string
-        Selected user's Auth0 id
-    project_list: array
-        List of projects to allocate
-    """
-    print(f"Check whether the user is in the DB, if not, add the person to the DB")
-    if not _is_user_in_db(user_id):
-        print(f"{user_id} is not in the DB so updating it.")
-        _add_user_to_db(user_id)
-        db.session.flush()
-
-    for project in project_list:
-        _add_user_project_to_db(user_id, project["value"])
-
-
-def remove_projects_from_user(user_id, project_list):
-    """Remove projects from the chosen user
-
-    Parameters
-    ----------
-    user_id: string
-        Selected user's Auth0 id
-    project_list: array
-        List of projects to remove from the DB
-    """
-    for project in project_list:
-        _remove_user_projects_from_db(user_id, project["value"])
-
-
-def _get_user_permissions(requested_user_id):
-    """Retrieve all permissions for the specified user
-
-    Returns
-    -------
-    A list of permission names
-    """
-    all_allowed_permission_for_a_user_list = models.UserPermission.query.filter_by(
-        user_id=requested_user_id
-    ).all()
-
-    return [
-        permission.permission_name
-        for permission in all_allowed_permission_for_a_user_list
-    ]
+    return users_projects_dict
 
 
 def get_all_users_permissions():
@@ -341,15 +301,50 @@ def get_all_users_permissions():
               permission_name: [permission_name]
             }
     """
-    # Get all allowed permission from the DB.
-    all_allowed_permission_list = models.UserPermission.query.all()
+    # Get all assigned permission from the DB.
+    all_users_permissions_list = models.UserPermission.query.all()
 
-    allowed_permission_dict = defaultdict(list)
+    users_permissions_dict = defaultdict(list)
 
-    for permission in all_allowed_permission_list:
-        allowed_permission_dict[permission.user_id].append(permission.permission_name)
+    for permission in all_users_permissions_list:
+        users_permissions_dict[permission.user_id].append(permission.permission_name)
 
-    return allowed_permission_dict
+    return users_permissions_dict
+
+
+def allocate_projects_to_user(user_id, project_list):
+    """Give user a permission of the chosen projects
+
+    Parameters
+    ----------
+    user_id: string
+        Selected user's Auth0 id
+    project_list: array of dictionaries
+        List of projects(dictionary) to allocate in the form of
+        [{label: project_name, value: project_id}]
+    """
+    print(f"Check whether the user is in the DB, if not, add the person to the DB")
+    if not _is_user_in_db(user_id):
+        print(f"{user_id} is not in the DB so updating it.")
+        _insert_user(user_id)
+        db.session.flush()
+
+    for project in project_list:
+        _insert_user_project_permission(user_id, project["value"])
+
+
+def remove_projects_from_user(user_id, project_list):
+    """Remove projects from the chosen user
+
+    Parameters
+    ----------
+    user_id: string
+        Selected user's Auth0 id
+    project_list: array
+        List of projects to remove from the DB
+    """
+    for project in project_list:
+        _remove_user_project_permission(user_id, project["value"])
 
 
 def get_all_permissions_for_dashboard():
@@ -359,61 +354,36 @@ def get_all_permissions_for_dashboard():
     -------
     A list of permission names
     """
-    # Get all allowed permission from the DB.
+    # Get all assigned permission from the DB.
     all_permission_list = models.Auth0Permission.query.all()
 
     return [permission.permission_name for permission in all_permission_list]
 
 
-def update_user_permissions(user_id, permission_list):
-    """Update/Insert user's allowed permission to a table,
-    Users_Permissions
-    
-    Parameters
-    ----------
-    user_id: string
-        Auth0's unique user id
-    permission_list: list
-        List of permission that user has. (From Auth0, trusted source)
-    """
-    # Sync the allowed_permission table to token's permission (trusted source)
-    # first before we check/update
-    _sync_permissions(user_id, permission_list)
-
-    print(f"Check whether the user is in the DB, if not, add the person to the DB")
-    if not _is_user_in_db(user_id):
-        print(f"{user_id} is not in the DB so updating it.")
-        _add_user_to_db(user_id)
-        db.session.flush()
-
-    for permission in permission_list:
-        _add_user_permission_to_db(user_id, permission)
-
-
 def _sync_permissions(user_id, trusted_permission_list):
     """Syncs the user access permissions with the ones from
     the token (the one source of truth)
-    Filter the allowed_permission table to remove outdated permission
+    Filter the users_permissions table to remove outdated permission
 
-    If the access token has the permission of A, B, C but allowed_permission table
+    If the access token has the permission of A, B, C but users_permissions table
     has the permission of A, B, C, D. Then remove the permission D from the
-    allowed_permission table as the token is the trusted source of permission.
+    users_permissions table as the token is the trusted source of permission.
 
     Parameters
     ----------
     user_id: string
         It will be used to find a list of permission with the function,
-        _get_all_allowed_permission_for_a_user(user_id)
+        _get_all_users_permissions_for_a_user(user_id)
     trusted_permission_list: list
-        The list to be compared with the list of permission from allowed_permission
+        The list to be compared with the list of permission from users_permissions
         table
     """
-    # Get a list of permission that are allocated to this user
-    unfiltered_allowed_permission_list = _get_user_permissions(user_id)
+    # Get a list of permission that are assigned to this user
+    unfiltered_assigned_permission_list = _get_user_access_permission(user_id)
 
-    for permission in unfiltered_allowed_permission_list:
+    for permission in unfiltered_assigned_permission_list:
         if permission not in trusted_permission_list:
-            _remove_user_permission_from_db(user_id, permission)
+            _remove_user_access_permission(user_id, permission)
 
 
 def write_request_details(user_id, action, query_dict):
@@ -431,7 +401,7 @@ def write_request_details(user_id, action, query_dict):
         E.g., Attribute -> Station
               value -> CCCC
     """
-    # Add to History table
+    # Insert to History table
     new_history = models.History(user_id, action)
     db.session.add(new_history)
     db.session.commit()
@@ -459,3 +429,27 @@ def write_request_details(user_id, action, query_dict):
             db.session.add(new_history)
 
     db.session.commit()
+
+
+def get_projects(access_level=None):
+    """Get all projects that have a given access_level from the Project table
+
+    Parameters
+    ----------
+    access_level: string
+        Determining the access level
+
+    Returns
+    -------
+    Dictionary in the form of
+    {
+        project_id: project_name
+    }
+    """
+    projects = (
+        models.Project.query.filter_by(access_level=access_level).all()
+        if access_level is not None
+        else models.Project.query.all()
+    )
+
+    return {project.project_id: project.project_name for project in projects}