+++ middleware/middleware/models.py
@@ -9,7 +9,10 @@ class UserProject(db.Model):
         "user_id", db.String(100), db.ForeignKey("user.user_id"), primary_key=True,
     )
     project_id = db.Column(
-        "project_id", db.Integer, db.ForeignKey("project.project_id"), primary_key=True,
+        "project_id",
+        db.String(100),
+        db.ForeignKey("project.project_id"),
+        primary_key=True,
     )
 
     user = db.relationship("User", back_populates="projects")
@@ -21,16 +24,19 @@ class UserProject(db.Model):
 
 
 class Project(db.Model):
-    project_id = db.Column(db.Integer, primary_key=True)
-    project_name = db.Column(db.String(100))
+    project_id = db.Column(db.String(100), primary_key=True)
+    project_name = db.Column(db.String(100), nullable=False)
+    access_level = db.Column(db.String(100))
 
     users = db.relationship("UserProject", back_populates="project")
 
-    def __init__(self, name):
-        self.project_name = name
+    def __init__(self, project_id, project_name, access_level="private"):
+        self.project_id = project_id
+        self.project_name = project_name
+        self.access_level = access_level
 
     def __repr__(self):
-        return "<Project %r>" % self.project_name
+        return "<Project %r>" % self.project_id
 
 
 class UserPermission(db.Model):